langsmith-tutorial/pyproject.toml
[project]
name = "langsmith-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.73.1",
    "langsmith>=0.7.33",
]
